refactor(app-extensions): Move reusable socket stuff to own directory
Web socket stuff is not only needed for notifications, but, for instance,
also in the devcon for the application log and the SQL log.

Refs: TOCDEV-3951
