+++ packages/app-extensions/src/notification/modules/socket/sagas.js
@@ -1,13 +1,13 @@
-import {takeEvery, all, call, take, put, select} from 'redux-saga/effects'
-import {eventChannel} from 'redux-saga'
-import {originId as originIdHelper, consoleLogger} from 'tocco-util'
+import {takeEvery, all, call, put, select} from 'redux-saga/effects'
+import {originId as originIdHelper} from 'tocco-util'
 
-import {getSocketUrl, notificationToToaster, TOASTER_KEY_PREFIX} from './socket'
+import {notificationToToaster, TOASTER_KEY_PREFIX} from './socket'
 import * as actions from './actions'
 import * as toasterActions from '../toaster/actions'
 import {toaster} from '../toaster/actions'
 import {notificationTransform} from '../../api'
 import {updateNotification, updateUnreadNotification, markAsRead} from '../center/actions'
+import socket from '../../../socket'
 
 export const notificationSocketSelector = state => state.notification.socket
 export const toastersSelector = state => state.notification.toaster.toasters
@@ -17,39 +17,18 @@ export default function* sagas(accept) {
     yield all([
       takeEvery(actions.SOCKET_MESSAGE_RECEIVED, messageReceived),
       takeEvery(toasterActions.REMOVE_TOASTER, toasterRemoved),
-      takeEvery(actions.CONNECT_SOCKET, connectSocket),
-      call(connectSocket)
+      call(initSocket)
     ])
   }
 }
 
-const websocketInitChannel = url =>
-  eventChannel(emitter => {
-    const ws = new WebSocket(url)
-
-    ws.onopen = e => emitter(actions.socketConnected())
-    ws.onmessage = e => {
-      const data = JSON.parse(e.data)
-      emitter(actions.socketMessageReceived(data))
-    }
-    ws.onclose = e => emitter(actions.connectSocket())
-    ws.onerror = err => {
-      consoleLogger.log('socket error', err)
-    }
-
-    return () => {}
-  })
-
-export function* connectSocket() {
+export function* initSocket() {
   const originId = yield call(originIdHelper.getOriginId)
   yield put(actions.setOriginId(originId))
-  const socketUrl = yield call(getSocketUrl)
-  const channel = yield call(websocketInitChannel, socketUrl)
-
-  while (true) {
-    const action = yield take(channel)
-    yield put(action)
-  }
+  yield call(socket.connectSocket, {
+    name: 'notification',
+    messageReceivedAction: actions.socketMessageReceived
+  })
 }
 
 export function* messageReceived({payload: {data}}) {
