+++ packages/app-extensions/src/socket/index.js
@@ -1,0 +1,5 @@
+import {connectSocket} from './socket'
+
+export default {
+  connectSocket
+}
